Move image processing stuff into crop
