+++ monsterbook/src/lib.rs
@@ -1,0 +1,4 @@
+extern crate image;
+extern crate imageproc;
+
+pub mod crop;
